Loading approach, storage data. Fixed #49
Fixed a bug that occurred when loading albums and was related to the track queue. During the fixing process came to the conclusion that it is necessary to change the approach of receiving data to the storage, separating the receipt for each entity. To do this, made changes on the server, replaced the method for getting tracks from the Playlist entity (now occurs in the Track entity) and added method for getting the album by artist id. Also made corrections in titles. Viewing components still need to be improved and optimized.