Queue and Play behavior fixes. Closed #52
Added Play button to the AlbumView and artistView components, now the queue fills with tracks only when you click on them, not when they are downloaded as it was before. Now pressing Play button in the Player starts playing the queue if the currentTrack is not selected. The Play button in the TrackView is now active. Added a placeholder for the queue. Improved custom hook for setting tracks into the queue. Now if the queue already has tracks, an attempt to set will only add tracks that are not present in queue. Added appropriate action for this. Added function to compare tracks in queue and component for correct display of Play/Pause buttons.